==> Experiment/EDA/01_Cleaning_Formatting.ipynb <==
import numpy as np
import pandas as pd

from datetime import datetime as dt
import random
import re

from tqdm import tqdm

from itertools import chain
from collections import Counter

data = pd.read_parquet("../../Data/split_4.parquet")

data.info()

data.sample(6)

data.StepName.value_counts()

past_review_steps = [
        "Phone Interview",
        "In Person Interview1",
        "Background Check",
        "Recruiter Reviewed",
        "Hire",
        "In Person Interview",
        "Phone Screen / Interview",
        "In Person Interview 2",
        "Background Check (Criminal & MVR)",
        "Background Ck - Crim & MVR",
        "Contingent Offer",
        "1st Interview",
        "Reference Interview",
        "Post Offer / Physical & Drug Screen",
        "Post-Offer Phys and/or Drug Screen",
        "In Person Interview2",
        "In Person Interview #1",
        "Extend Contingent Offer",
        "In Person Interview #2"
    ]
def categorize_step(step_name, past_review_steps=past_review_steps):
    # List of steps that signify the candidate made it past the resume stage
    return 1 if step_name in past_review_steps else 0

# Apply the function to your DataFrame
data['Step_Category'] = data['StepName'].apply(categorize_step)


data['Step_Category'].value_counts()

### We categorized the steps since we are interested in who is 'qualified' for the job, since performance on interviews (ultimately what leads the candidate to the hire) is not something we can predict from the resume.
- furthermore, this provides us more data than just looking at hires.

data.info()

data.dtypes

---

### Fields related to Opportunities
- OpportunityId
- ExternalBriefDescription
- ExternalDescription
- Title
- JobCategoryName

### Fields related to Applications
- ApplicationId
- IsRejected
- IsCandidateInternal
- pass_first_step
- Tag
- WorkExperiences
- Educations
- LicenseAndCertifications
- Skills
- Motivations
- Behaviors
- StepId
- StepName
- StepGroup

### Fields related to Criteria
- BehaviorCriteria
- MotivationCriteria
- EducationCriteria
- LicenseAndCertificationCriteria
- SkillCriteria

criteria_columns = ["BehaviorCriteria", "MotivationCriteria",
    "EducationCriteria", "LicenseAndCertificationCriteria",
    "SkillCriteria",]

for column in criteria_columns:
    print(data[column][66159]) # Random sample

data.drop(columns=criteria_columns, inplace=True)

data.columns

### The criteria could increase accuracy in the model, but it's not practical yet given the inputs we expect from a user. We'll drop criteria columns for now.

applicant_columns = ['ApplicationId', 'IsRejected','IsCandidateInternal',
                    'pass_first_step', 'Tag', 'WorkExperiences',
                    'Educations', 'LicenseAndCertifications', 'Skills',
                    'Motivations', 'Behaviors', 'StepId', 'StepName',
                    'StepGroup']

opportunity_columns = ["OpportunityId",
    "ExternalBriefDescription",
    "ExternalDescription",
    "Title",
    "JobCategoryName"]

data['WorkExperiences'][85974]

### Extracting the job titles and work experiences from applications

data['Skills'][66159]

def extract_details(entry):
    job_titles, job_responsibilities, applicant_educations, skills = [], [], [], []
    
    # Extract Job Details
    work_experience = entry.get('WorkExperiences', [])
    for job in work_experience:
        title = job.get('JobTitle', '')
        responsibility = job.get('WhatYouDid', '')
        if title: job_titles.append(title)
        if responsibility: job_responsibilities.append(responsibility)
    
    # Extract Education Details
    education_list = entry.get('Educations', [])
    for edu in education_list:
        degree = edu.get('Degree', '')
        if degree: applicant_educations.append(degree)
    
    # Extract Skills
    skill_list = entry.get('Skills', [])
    for skill in skill_list:
        skill_name = skill.get('Skill', '')
        if skill_name: skills.append(skill_name)

    return "++".join(job_titles), "++".join(job_responsibilities), "++".join(applicant_educations), "++".join(skills)

# Apply the function and create new columns
data[['Applicant_Job_Titles', 'Applicant_Job_Responsibilities', 'Applicant_Education', 'Applicant_Reported_Skills']] = data.apply(lambda x: pd.Series(extract_details(x)), axis=1)

data[['Applicant_Job_Titles', 'Applicant_Job_Responsibilities', 'Applicant_Education', 'Applicant_Reported_Skills']]

### Since we will be cleaning up punctuation marks, I separated values using "++", which is unlikely to be used in the data. Therefore, we can still separate the values later on.

data.columns

data.sample(3)

data.columns

cols_to_drop = ['WorkExperiences', 'Educations', 'Skills', "JobCategoryName",
                'IsCandidateInternal', 'WorkExperiences', 'Educations',
                'LicenseAndCertifications', 'Skills', 'Motivations', "IsRejected",'Behaviors',
                'StepId', 'StepName', 'Tag', 'StepGroup',]
data.drop(columns = cols_to_drop, inplace=True)

data

data.to_parquet("../../Data/split_4_cleaned.parquet")

### Here, we cleaned the application data by removing columns we won't need for the model.
- People interview for roles they're qualified for, and that's typically based on their skills, previous experience, and education.
- Therefore, we've taken applicant data such as job titles, experiences, skills, and education.
- We have held onto opportunity columns that provide the requirements for the role (ExternalDescription and ExternalBriefDescription) as well as job title (so we can infer some skills necessary for the role, if we choose to)
- We also have identifying information (OpportunityId and ApplicationId) as unique identifiers
- To create a supervised learning algorithm, we also have a variable which we can use to see whether or not the applicant made it to the next steps, such as phone screenings and interviews.
