==> Experiment/EDA/02_Cleaning_Text.ipynb <==
# Cleaning the text across the applications

import pandas as pd
import numpy as np

import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
import warnings

# Ignore ipykernel warning
warnings.filterwarnings('ignore', category=DeprecationWarning, module='ipykernel')

applications = pd.read_parquet('../../Data/split_4_cleaned.parquet')

applications.columns

# Change column names
applications.columns = ['opportunity_id', 'application_id', 'opportunity_brief_description',
            'opportunity_description', 'opportunity_title', 'application_pass_first_step', 'application_step_category',
            'application_job_titles', 'application_job_responsibilities',
            'application_education', 'application_reported_skills']
# Remove duplicates for job_descriptions
original_descriptions = applications.opportunity_description.unique()

#These are all lists of strings
cols_not_to_concat = ['application_id', 'application_pass_first_step', 'application_step_category']
application_cols = [col for col in applications.columns if 'application_' in col and col not in cols_not_to_concat]
opportunity_cols = [col for col in applications.columns if 'opportunity_' in col and col != 'opportunity_id']

applications['application_concat'] = applications[application_cols].astype(str).agg('--'.join,axis=1)

applications['application_full_tokenized'] = applications['application_concat'].apply(nltk.word_tokenize)

application_cols.append("application_concat")
application_cols.append("application_full_tokenized")

#If column is object, it will become string
for column in applications.columns:
    if applications[column].dtype == object:
        applications[column] = applications[column].astype("string")

applications.info()

## Preprocessing the data

applications.columns

columns_to_preprocess = ['opportunity_brief_description', 'opportunity_description', 'opportunity_title',
                        'application_job_titles', 'application_job_responsibilities',
                        'application_concat', 'application_full_tokenized']

import nltk
nltk.download('stopwords')
nltk.download('wordnet')
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import RegexpTokenizer

tokenizer = RegexpTokenizer(r'\b\w+\b')
stop_words = set(stopwords.words('english'))
lemmatizer = WordNetLemmatizer()

def strip_html_tags(text):
    html_pattern = re.compile('<.*?>')
    text = re.sub(html_pattern, '', text)
    return text

def preprocess_text(text):
    # Strip HTML tags and convert to lowercase
    text = strip_html_tags(text).lower()
    
    # Remove unwanted strings using compiled regex pattern
    # This changed
    unwanted_strings = ['sign on bonus', 'sign bonus',
                         'full time', 'part time',
                         'day shift', 'night shift', 'second shift',
                         'third shift', 'first shift', 'none', 'nbsp', 'amp']
    unwanted_pattern = re.compile('|'.join(map(re.escape, unwanted_strings)))
    text = re.sub(unwanted_pattern, '', text)

    # Tokenize the text and remove stopwords while lemmatizing
    tokens = [lemmatizer.lemmatize(token) for token in tokenizer.tokenize(text) if token not in stop_words]

    # Join the tokens back into a single string
    text = ' '.join(tokens)

    return text

for column in columns_to_preprocess:
    print(f"working on {column}")
    applications[column] = applications[column].apply(preprocess_text)

applications = applications.astype(str)

# Re-run the test code to see if the data types have changed
for col in application_cols:
    print("Column:", col)
    dtypes_of_items = set()  # Use a set to store unique data types
    for cell in applications[col]:
        for item in cell:
            if item is not None:
                dtypes_of_items.add(type(item))
            else:
                dtypes_of_items.add(type(None))
    print("Unique data types:", dtypes_of_items)

### Everything is a string, which is essential for compatability with the model

applications.sample(3)

### We want zero NoneType

applications.to_parquet('../../Data/split_4_preprocessed.parquet')
